+++ app/client/event_listeners/listener_website_link.js
@@ -1,0 +1,20 @@
+
+// ┌─────────────────────────────────────┐
+// │                                     │
+// │          Streetview button          │
+// │                                     │
+// └─────────────────────────────────────┘
+
+export function listener_website_link(){
+
+    const websiteLink = document.getElementById("drawer-link")
+
+    websiteLink.addEventListener('click', async function(event) { 
+
+        event.preventDefault();
+        
+        // Open in chrome.
+        require("electron").shell.openExternal(websiteLink.href);
+
+    });
+}